Implements major content changes, a nav bar, and general cleanup and refactoring
- Adds a navigation bar for the application

- Touches up the content in various areas and introduces the "bnny" page

- Removes the lightweight DnD application to take a different direction with content

- Substantial styling updates

- Miscellaneous code cleanup and refactoring
