+++ README.md
@@ -1,10 +1,8 @@
 # Bnny
 
-This is a simple web application that I'm working on as a fun side project.
+This is a simple web application that I am working on as a fun side project. You can find the live version at [bnny.us](https://bnny.us)!
 
-License information for this software can be found in [`LICENSE`](LICENSE). All of the other projects referenced in the build steps are licensed under an MIT license as well.
-
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+This software is licensed under an MIT license. Specifics can be found in [`LICENSE`](LICENSE). This project was originally bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
 ## Available Scripts
 
@@ -15,13 +13,3 @@ This project was bootstrapped with [Create React App](https://github.com/faceboo
 - Launch the test runner with `npm test`
 
 - Run the application's linter with `npm run lint`
-
-## Building Bnny
-
-While you can simply build the application using the node configuration `npm run build`, there are some sub applications you will need to setup first.
-
-Simply navigate to the `public` directory and clone with following repositories into the following directories.
-
-- `https://github.com/5etools-mirror-3/5etools-src` -> `5eTools`
-
-Once they are cloned into the correct directories, run `npm run build`, and presto! Copy the build directory into anywhere you can host a website.
