+++ public/index.html
@@ -15,7 +15,7 @@
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>Joy</title>
+    <title>Bnny</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
