+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Joy Website",
-  "name": "Joy's Website",
+  "short_name": "Bnny Website",
+  "name": "Bnny Website",
   "icons": [
     {
       "src": "bnyy.png",
