+++ src/Components/Pages/Home.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Home = () => {
+
+    return(
+    <div className="App">
+      <header className="App-header">
+        <p>
+          This is a fun little work-in-progress project. It is still under heavy renovation, so stay tuned!
+        </p>
+      </header>
+    </div>);
+  };
+  
+export default Home;
